Implement move-result-pseudo instructions [1/N]
Summary:
Background
==========

Opcodes that write to a register (say v0) but may also throw are somewhat
tricky to handle. Our dataflow analyses must consider v0 to be written only
if the opcode does not end up throwing.

For example, say we have the following Dex code:

  // start try block
  const/4 v0 #0
  sget-object v0 <some field of type LFoo;>
  return-void
  // end try block

  // exception handler
  invoke-static {v0} LQux;.a(LFoo;)V

If `sget-object` throws, it will not have written to v0, so the `const/4` is
necessary to ensure that v0 is always initialized when control flow reaches
B2. In other words, v0 must be live-out at const/4.

Prior to this diff, we dealt with this by putting any potentially throwing
opcodes in the subsequent basic block when converting from Dex to IR:

  B0: <throws to B2>
    const/4 v0 #0
  B1:
    sget-object v0 <some field of type LFoo;>
    return-void
  B2: <catches exceptions from B0>
    invoke-static {v0} LQux;.a(LFoo;)V

This way, straightforward liveness analysis will consider v0 to be live-out at
const/4. Obviously, this is still somewhat inaccurate: we end up considering
some values as live when they are really dead. Being conservative about
liveness generally doesn't create wrong behavior, but turns out to be
particularly inefficient when allocating check-cast.

To recap on check-cast: in our IR, it has both a src and a dest since it
changes the type of a register, but when we convert it to an DexInstruction
we need to insert a move-object instruction before the check-cast if the src
and dest are different registers. As such, we must be careful when allocating
its operands. For example, say we have the IR:

  // Assume that Foo is a subclass of Bar, and Baz is some unrelated type
  B0: <throws to B2>
    sget-object v0 <some field of type LBaz;>
    sget-object v1 <some field of type LFoo;>
  B1:
    check-cast v0, v1 LBar // this is an IRInstruction, so check-cast has a dest and a src
    invoke-static {v0} LQux;.a(LBar;)V
    return-void
  B2: <catches exceptions from B0>
    invoke-static {v0} LQux;.a(LBaz;)V

Observe that we _cannot_ coalesce v0 and v1 here, even though it superficially
appears that v0 is never written to while v1 is live. If we did that, we would
have (in Dex code):

  sget-object v0 <some field of type LBaz;>
  sget-object v1 <some field of type LFoo;>
  move-object v0, v1
  check-cast v0 LBar;
  invoke-static {v0} LQux;.a(LBar;)V
  return-void
  // exception handler -- oh no, v0 has type Foo here, which is a type error!
  invoke-static {v0} LQux;.a(LBaz;)V

To avoid this, when allocating check-cast, we look at the previous opcode and
avoid coalescing v0 with any register that's live-out there, since that means
that some successor of B0 might need to use that value. So in the above
example, we do not coalesce v0 and v1 because v1 is live-out at B0.

This is a correct strategy, but with a catch... since check-cast is in B1,
however, v0 is *always* live-out at B0, regardless of the code in B2. So we
never coalesce v0 and v1 with this approach. Now we could hack around this by
looking at all the successor of B0 when allocating our check-cast and seeing if
it actually uses v1, but move-result-pseudo seems to be a cleaner solution:

  B0: <throws to B2>
    sget-object v0 <some field of type LBaz;>
    sget-object v1 <some field of type LFoo;>
    check-cast v1 LBar;
  B1:
    move-result-pseudo-object v0
    invoke-static {v0} LQux;.a(LBar;)V
    return-void
  B2: <catches exceptions from B0>
    ...

Now v1 is live-out at B0 if and only if it is live-in at B2. This allows us
to coalesce a whole lot more of check-cast operand registers, resulting in
considerable space savings.

Specific Changes
================

* All dex opcodes that have a dest register and may also throw an exception are
  represented in our IR by two instructions: one piece that may throw but does
  not write to a dest, and one move-result-pseudo instruction that writes to a
  dest but does not throw.

* IRCode's remove/replace methods will always remove these opcodes as a pair.
  We don't yet enforce inserting these opcodes as pairs, though we really
  should.

* MFLOW_FALLTHROUGH items no longer have a throwing_mie entry.

* RegAlloc (aside from its live range splitting code) and the typechecker have
  been updated to use the new IR correctly. The other passes are *not* updated
  in this diff and are broken if you are at this specific revision; I am
  splitting out those (mostly mechanical) changes since this diff is pretty
  large already. I will stack land them.

Reviewed By: justinjhendrick, arnaudvenet, maxmg22

Differential Revision: D5995657

fbshipit-source-id: 66430df2abf0786979b398e3b31e17b1830731f7