localdce: Adapt to move-result-pseudo instructions [2/N]
Summary:
All dex opcodes that have a dest register and may also throw an exception are
now represented in our IR by two instructions: one piece that may throw but
does not write to a dest, and one move-result-pseudo instruction that writes to
a dest but does not throw.

This diff makes localDCE compatible with the new IR.

See {D5995657} for more background.

Reviewed By: grfrost, maxmg22

Differential Revision: D5995704

fbshipit-source-id: a777a74ddf9a4e4ccc74c9ceb2497837a2bbcd55