simpleinline: Adapt to move-result-pseudo instructions [4/N]
Summary:
All dex opcodes that have a dest register and may also throw an exception are
now represented in our IR by two instructions: one piece that may throw but
does not write to a dest, and one move-result-pseudo instruction that writes to
a dest but does not throw.

Since we use opcode count as a heuristic for inlining, in order to avoid
changing behavior, we should ignore these extra move-result-pseudo opcodes in
our counts.

See {D5995657} for more background.

Reviewed By: grfrost, maxmg22

Differential Revision: D5995742

fbshipit-source-id: 1ffa6802789552eb40e2a785ba448cb22a5cb3d6