peephole: Adapt to move-result-pseudo instructions [5/N]
Summary:
This was a pain... but mostly mechanical transformations...

See {D5995657} for more details.

Unfortunately this is not a 100% reproduction of the pre-IR-refactor behavior
because of the way the algorithm works -- when it encounters a failure to match
beyond the second opcode of the pattern, it skips the entire length of the
pattern. The pseudo opcodes mean that some 1-opcode partial matches are now
2-opcode partial matches, causing more skippage.

Reviewed By: justinjhendrick, maxmg22

Differential Revision: D5995832

fbshipit-source-id: dbdf0160a02d19e3d0dc6c775fda3d5bf86ebf0e