More compact encoding of literals in IRInstruction
Summary:
Instead of storing m_literal as a separate field, we can store it in the same
tagged union as the field/method/type etc fields, since any given opcode only
contains one member of that union.  Encoding it this way makes writing a parser
for IRInstruction cleaner (upcoming diff). It also should give us some memory
savings.

For uniformity with the naming conventions of other methods, this diff renames
`literal()` to `get_literal()`... I realized belatedly that this makes the diff
a little messy to review, oops

Reviewed By: arnaudvenet

Differential Revision: D6039688

fbshipit-source-id: c506e04b2c2b1e3e3da794798eb5c978824537ab