Fix buck name collisions in extracted_res_deps
Summary:
The addition to `redex_instr_test` marco in D5978862 has a mistake where if multiple tests are defined that depend on a given resource file, multiple buck genrules will be defined with the same name. This throws a parsing error in Buck.

Fix it by making the genrule names unique.

Differential Revision: D6056255

fbshipit-source-id: c313406240172673c1ee4c41d0f7bfb3a105f0f2