Begin implementing IR assembler
Summary:
So I don't think `dasm()` was really cutting it as a way to create instructions
(still too verbose). Moreover, I realized that I was avoiding writing unit tests for
regalloc because they were so painful to write. As such, I decided to work on a
de/serialization format, building off arnaudvenet's s-expression library. The reasons in
detail:

* More compact than any embedded DSL
* Serialization opens the possibility of saving/reloading IR between passes
* Serialization provides an easy way to compare equality of a structure that
  has a lot of internal pointers (IRCode) -- just check if the serialized output
  is the same
* Will be easier to change the IR when our tests use an abstraction of it

This diff just sketches out the beginnings of the deserializer. It only handles
a subset of instructions (just those needed for the regalloc unit test). It
does provide a nice way to express labels and jumps, but it does not support
try-catch blocks yet.

Also, presently, the serialization format doesn't really take advantage of
S-expressions' ability to encode trees (it's just a linear list of instructions
right now) -- but S-exprs will make it easy to express these lists as child
expressions of methods and classes.

Reviewed By: arnaudvenet, justinjhendrick

Differential Revision: D6039718

fbshipit-source-id: 0d47ce11a5a897805e73769adcf49b551a0c05bb