Begin implementing IR disassembly
Summary:
{D6039718} mentioned that "Serialization provides an easy way to compare
equality of a structure that has a lot of internal pointers (IRCode) -- just
check if the serialized output is the same". This diff starts to implement
deserialization for this purpose.

Like the previous assembler diff, this only implements serialization for a
subset of instructions, and does not support try-catch blocks either. I'll
follow up on that soon...

While at it, I added support for comments to the S-expression parser. It makes
unit tests a little nicer to write.

Reviewed By: arnaudvenet

Differential Revision: D6056260

fbshipit-source-id: 3b8bb6fb40ad21551515a65b2ce6299644017592