Fix invalid read in IRInstruction::hash
Summary:
Wow, I love ASAN now. I never would have found this bug otherwise.
DexOpcodeData has a `size` method that inherits from `DexInstruction`.
The `size` method returns the size of the _whole instruction_, not the
size of the data segment. In IRInstruction::hash, we want the size of
the data segment, not the instruction

Reviewed By: emma0303, int3

Differential Revision: D6081904

fbshipit-source-id: 734fb3f01ab0ec91e8805d9be6121333a668b3cf