Fixes for redex-all/main.cpp (1/2)
Summary:
We now have a strong need to build Redex natively on Windows (i.e., using Visual Studio 2015/2017, a.k.a. MSVC). It will require some changes in Unix-only/non-portable code.

This is the first patch for Redex on Windows.

Here are what I did:
- Install Visual Studio 2017 with Visual Studio 2015 tool chain. (VS allows you to switch compiler versions easily.)
- Clone the redex repository.
- Create an empty command-line C++ project.
- Add all source code as shown in CMakeList.txt (e.g., libredex/*.cpp).
- Add -I options appropriately.
- Switch to a x64 build configuration. (We don't support x86.)
- Go to NuGet package manager and install (1) boost v1.65.1 and (2) jsoncpp v0.6.0.1.
- First, just compile `redex-all/main.cpp` only. Building the entire project would expose a ton of errors. So let's attack the `main.cpp` first.
- Fix the errors.

This diff will make it compile `main.cpp` successfully, except for `parse_args` that requires `<getopt.h>`.

NuGet does have a "GetOpt" library. But it's quite stale; not compilable by MSVC 14.0 (2015) and 14.1 (2017). It's better to rewrite the `getopt_long` part by `boost::program_options`, which will be the follow-up diff.

One of the goals for this work is minimizing ugly `#ifdef` in the code. They should be mostly shown in the headers.

Reviewed By: justinjhendrick

Differential Revision: D6087133

fbshipit-source-id: f7fe8c908fa58269a27c5dcb47da43597bb964c0