assembler: actually print labels
Summary:
So I had actually forgotten to print the labels, rendering the unit
tests a bit sketchy... I've fixed that and added a unit test for the
disassembler itself so we can actually see what it's printing.

I also changed the assembler syntax a bit -- I don't think labels will
ever be a compound structure (at least, if we are using smali syntax
as inspiration) -- so I've changed them from lists to atoms.

Reviewed By: maxmg22, arnaudvenet

Differential Revision: D6131326

fbshipit-source-id: b9c73061bd48596e04224d9a7cd936a241fbcc06