Static analysis that identifies access paths into immutable structures
Summary: This analysis infers the access paths (in terms of sequences of getters) into an immutable data structure passed as an argument to a method. For instance, the analysis can tell that at a certain point in the code, register `v_n` points to the subcomponent `p1.getA().getB(C).getC()` of the structure contained in parameter #1.

Reviewed By: justinjhendrick

Differential Revision: D6096746

fbshipit-source-id: 9f67922d124acb8465722ed4b6e77935962cc23d