Resolve method collision on signature update
Summary:
When updating method signatures, two methods could ended up sharing the exact same signature. That is the only difference in their signature was the erased type.
This diff introduces the handling of colliding methods when performing method signature update.

The only colliding methods we found so far are 3 ctors. The current solution is to append an additional parameter to the ctor and patch its call site to pass in a dummy const int of value 42.

Reviewed By: emma0303

Differential Revision: D6118605

fbshipit-source-id: e890c22b6e9d025ebf16c6e4d6fd71031813b6e9