Better representation of multiple edges between blocks in CFG
Summary:
Previously, if we had multiple paths of control flow between the same
two blocks, we would record one edge in the CFG while setting multiple edge
flags. However, this was not a very natural abstraction for the
MonotonicFixpointIterator, which expected one distinct edge per control flow
path. This diff refactors the CFG to use that representation instead. Edge
flags are now a plain enum attached to each edge (instead of a bitset
describing the types of all possible edges between two blocks).

Specifically, the problem was with the following code in constprop:

  (load-param v0)
  (if-eqz v0 :next) ; This jumps to the next opcode regardless of whether
                    ; the test is true or false. So in this case we cannot
                    ; conclude that v0 == 0 in the 'true' block, since that
                    ; is identical to the 'false' block.
  :next
  (if-eqz v0 :end)
  (const/4 v0 1)
  :end
  (return-void)

Without multiple edges explicitly represented, the MonotonicFixpointIterator
was not calling join() on the domains from the multiple control flow paths
leading to the :next block, causing erroneous analysis results.

In addition to the CFG changes, this diff also changes the
MonotonicFixpointIterator to pass the edge object to analyze_edge. This allows
the implementors of the fp iterator to incorporate information about the edge
(such as whether it jumps to the true or false branch) in their analyses.

With the addition of edge src/edge target, MonotonicFixpointIterator needs quite
a lot of accessors now. To simplify the code, and to make it easy to do backwards
iteration, we bundle these accessors into a single interface and add a backwards
iteration adaptor.

Reviewed By: justinjhendrick, arnaudvenet

Differential Revision: D6208017

fbshipit-source-id: 438c2d164ba2b156b6faa1550798d201f53dcc4b