Remove unused @lint-avoid-* from (mostly) non-fbcode locations
Summary: This diff is like D6215811 and D6232366, but for `fbandroid/`, `fbobjc/` and `xplat/`. These changes are were done independently of the `fbcode/` ones because they were so few in number and it made landing the fbcode diff easier. EDIT: Added one fbcode location that snuck in after I landed D6232366.

Reviewed By: mzlee, maxmg22

Differential Revision: D6233351

fbshipit-source-id: f675c8488ce9e0138a55c823e38b4bb5b8d13e4a