constprop: Assume abstract registers
Summary:
See the summary of {D6241960} for context.

Note that this diff does not cause constprop to generate
non-register-width-respecting bytecode. If the input bytecode does give wide
registesr two IDs, the output bytecode will still respect this invariant. This
diff just simplifies our analysis, and allows constprop to keep working even if
the input does not respect register widths. This allows for a gradual migration
plan.

Reviewed By: arnaudvenet

Differential Revision: D6241966

fbshipit-source-id: 433adf32eeae51c2158061763ec6f6b06a926d18