OAT V124 writing support in Oatmeal
Summary:
This is a very early review of the work done to get Oreo VDEX/ODEX generation.  Still a WIP but useful to get some early eyes.
Haven't tested on the device yet but getting closer.  Currently the generated files are structuraly very similar to the dex2oat
generated VDEX/ODEX pairs but there are a few rough edges that have to be ironed out:

1. Get the VDEX checksum right instead of writing garbage.
1. Modify the DexTricks code to instead of calling oatmeal with an output .oat file, create the directory where
the odex/vdex files should live and then invoke oatmeal passing that directory as the target.

Reviewed By: mlogan

Differential Revision: D6204324

fbshipit-source-id: fdc321fbee0ceaddbc38938ac84849f480c84ada