Make SingleImpl aware of multiple DexStore
Summary:
SingleImpl was not honoring cross DexStores constraints and allowing implementation to leak in DexStores that were loaded before the DexStore with the implementation was loaded. Thhus causing a class not found exception.
This checks that when an optimization is taking place the implementation is not in a DexStore after the interface definition.

build[automation_instagram_,automation_fbandroid_,automation_messenger_ics_redex_,automation_messenger_art_redex_]

Reviewed By: int3

Differential Revision: D6262153

fbshipit-source-id: cbfdc94870a93a6dd5b42eea3e023fe50dd4c884