Simplify Ctor dedup
Summary:
This is a reland of D6283330, which was reverted due to a conflict with a test in redex-stable.

The logic around dedupping ctors in our code can be subsumed by the use of dedup helper.
Since the DedupCtor annotation does not imply any additional transformation, we don't really need to treat it specially.
This diff removes all the code around the DedupCtor annotation and relies on dedup helper to dedup when creating ctor dispatch.

build[automation_fbandroid_ics]

Reviewed By: emma0303

Differential Revision: D6322118

fbshipit-source-id: 529ed88b43bd371626c279bd07f324b1d5c729ab