sort fields before output
Summary: This diff is only really pertinent to the Obfuscator pass which is not currently enabled. DexOutput::generate_static_values generates sv's based on the order of the fields in the DexClass. If those fields are not ordered, then which field maps to which sv in the output Dex will be screwed up and cause VM aborts. So, we simply sort the dex fields before emitting static values in order to ensure that the sv's are ordered in a way that's reflective of the fields' order. This hasn't been an issue before because we've tended to not do much to change the order of static fields on classes.

Reviewed By: minjang

Differential Revision: D6329223

fbshipit-source-id: 90fb81250d627da97b484cc4e75f2e061f01b17f