Correctly abstract away the contents of the result register
Summary:
Consider the following piece of code:

  invoke-virtual ...
  move-result-object v1
  invoke-interface ...
  move-result-object v2

`ImmutableComponentAnalyzer` might assign an access path to register `v1` after the `invoke-virtual` operation. The `invoke-interface` operation is ignored by the analyzer. However, the contents of the result register must be clobbered, otherwise the next `move-result-object` instruction will assign the access path contained in `v1` to `v2`, which is incorrect.

Reviewed By: int3

Differential Revision: D6334734

fbshipit-source-id: 9cb0a1c3c1803d6f074271df769f8f76274191ff