Workaround that prevents stack overflows in WTO computation
Summary:
When the depth of recursion in the `visit()` method reaches a certain threshold, the rest of the computation is forked to a new thread.

This effectively splits a very long recursive call across multiple threads (that are executed in sequence), thus spreading the stack frames across multiple stack spaces. The purpose is to prevent a stack overflow when processing very deep control-flow graphs.

Reviewed By: minjang

Differential Revision: D6399233

fbshipit-source-id: 1bda6d2694c905cb9c5a95a7adf263107147561a