TypeSystem api to retrive the set of concrete methods
Summary:
We need a way to match a set of types against a virtual or interface scope.
Arnaud,
this may need more work in time particularly as API, we'll talk more.
However the idea is that you get either a VirtualScope from an invoke_virtual or an InterfaceScope from an invoke_interface. You should use the DexMethodRef that comes with the instruction.
Then we have the type set and we call TypeSystem::select_methods(scope, type_set) and you get back a set of concrete methods.
I think that is what we want but let me know otherwise.
You can take a look at the unit test in case it helps and I will improve that as I think more about this

Reviewed By: arnaudvenet

Differential Revision: D6408971

fbshipit-source-id: 0b55269313603c1429ba50028031b6f6bcb1d8e2