Oatmeal for Oreo
Summary:
After some battling with AOSP finally figured out what caused the ODEX/VDEX pairs to be rejected by the runtime.  There were minor mistkaes + a new necessary field in the KeyValue store
that states the OAT file supports concurrent-copying.  We had to remove the BSS section since the ART runtime tries to cache stuff there and fails.

I verified the generated ODEX/VDEX pairs now load, `dex2oat` no longer runs on the secondary dexes and all the loading of the ODEXes ends up in `kOatUpToDate` which means it worked.

Note: this set of changes only fix oatmeal to properly generate the binaries but we will have to make some dextrick changes to actually have oatmeal kick in properly.  Right now we have all that machinery disabled.  Expect a diff soon with that stuff to have this working end to end modulo a GK.

Reviewed By: mlogan

Differential Revision: D6445041

fbshipit-source-id: 12a9cfc8e6ced7d1b8a7acf45dce66f610656428