Fix the interface for the reduction operation in ReducedProductAbstractDomain
Summary:
Dynamic binding is disabled in constructors and hence, we can't define the reduction operation as a virtual method if we want to invoke it in the constructor of `ReducedProductAbstractDomain`. Making it a static method operating on a tuple of abstract values solves the issue.

I also added a call to `normalize()` at the end of the reduction, since there's no guarantee that `reduce()` will leave the tuple in a normalized state.

Reviewed By: int3

Differential Revision: D6462265

fbshipit-source-id: 7dd5c899b21f3cb98109db7998af59522d5ec191