Fix for the obfuscator
Summary:
when renaming virtuals we should account for names in derived types
that cannot be renamed and would introduce a collision (and a verification
error).
That happens when a method in a derived class is an interface implementation
and later on we find out we cannot rename that method. In such cases we may
have left a name that collides with something up the hierarchy that was
unrelated.
The fix is a bit more brutal and conservative than it should or could. It does not use any name that a
derived class may use. That is particularly bad when obfuscation is applied on
top of PG obfuscation but I guess that is a scenario we don't particularly
care.
It may also make the space of used names worse than it was already.
Oh well....

On a side note we seem to have cannot_rename on things that PG has renamed...
odd...

Reviewed By: newobj

Differential Revision: D6471186

fbshipit-source-id: 49042b5955916d3dd8feddbdceb9b4acd44495a2