Add uninstantiated in the root store mergeables at the end
Summary:
The bug here was that we are only checking in the root store if the class is
instantiated or not. And FrescoImage$Builder was instantiated in a different
one.

I wasn't correctly adding the non instantiated classes, since it would have
been skipped.

This would only affect classes defined in root store and used in a different
one. As is the case for FrescoImage$Builder

Reviewed By: dariorussi

Differential Revision: D6529804

fbshipit-source-id: 1e67e5bc67d66685d9dc37f4ee491874a43f8e05