kill @Signature annotations that reference non-existent types
Summary:
It seems that many Redex transforms don't clean up Signature annotations very well. It's not even always clear what should be done to a Signature that contains a type that we delete or erase/merge/singleimpl.

It seems reasonable to assume that  a Singleton which references a non-existent type is just as dangerous as no Signature - except no Signature is a lot smaller :) So by this logic, add a flag that kills Signature annotations that reference bogus classes.

Reviewed By: maxmg22, justinjhendrick

Differential Revision: D6540361

fbshipit-source-id: dc47aa720447a0bab76c6c91e30516b7eab62f77