Fix for obfuscation of miranda methods
Summary:
javac generates calls to method refs on a class that implements an interface but it does not defines the method e.g.
  interface I { m(); }
  class A implements I {}
  class B extends A() {}

  void somewhere() { B b = ...; b.m(); /* invoke_virtual B.m() */ }

while binding we want to get this into account.
Now the thing is that I believe we do take it into account on the scopes but not when resolving refs in opcodes.
I am not sure if this is showing elsewhere in other forms, I thought about it and convinced myself we can review all of that at another time.
That is, except for signature changes (name,proto) it does not matter if we are missing some association. I think at the end things will reconcile.

Of course, we need to review the binding story and make sure we have the right API.
Added one for now, used for the case described but we may do better.

Reviewed By: minjang

Differential Revision: D6637117

fbshipit-source-id: 01d0d6ef44cd5ba6ae16475957ceb18028d65dd3