Create IROpcode
Summary:
Previously, IRInstruction reused the same opcode type (DexOpcode) as DexInstruction, but gave different semantics to some of them. This was a potential source of confusion. Additionally, some DexOpcodes would only ever appear in a DexInstruction, and others would only ever appear in an IRInstruction, but there was no static enforcement of this property. Instead we had a bunch of needless runtime assertions. This refactor should make things simpler and clearer.

Concretely, this diff creates an IROpcode type that does not have any 2addr- or range-format opcodes. Additionally, it removes the load-param and move-result-pseudo opcodes from DexOpcode -- these only need to exist in IROpcode.

I plan to land a few more IROpcode simplifications after this diff. For example, const-, move-, and {i,s,a}{get,put} opcodes can be represented more parsimoniously in the IR.

Reviewed By: maxmg22

Differential Revision: D6637902

fbshipit-source-id: 12010e2ecd6f97bdf8ed558c512146f1a14eed86