Refactor ProguardMatcher.cpp for easier parallelization
Summary:
Do some refactoring so that we can easily parallelize `process_keep`.

The most noticeable change is replacing a single `regex_map` (hard to parallelize) by per-iteration local variable. It virtually doesn't hurt performance.

Added `const` as much as possible. The two `count`s in `MemberSpecification` and `KeepSpec` are marked as `mutable`. They are not critical; just for printing and stats purpose. That allows us to add more `const` to `MemberSpecification` and `KeepSpec`.

I also found that `hierarchy` was created multiple times. Confirmed that the contents of all `hierarchy` were the same.

Reviewed By: int3, justinjhendrick

Differential Revision: D6457938

fbshipit-source-id: ec4aeed2b41f2da74d694a5ffc3c3617b3768c0a