Flatten GlobalConstProp into IntraProcConstantPropagation
Summary: Originally, intraprocedural constant propagation was designed to operate on a generic control-flow structure. maxmg22 informed me that this was because it was difficult to build nontrivial CFGs of Redex's IR in unit testing; this genericity enabled him to write unit tests by running the constant propagation on a simple language, in which interesting control flow was easier to construct. However, now that we have the IRAssembler, it does not seem necessary to maintain that genericity (and the extra boilerplate code that it entails).

Reviewed By: justinjhendrick, arnaudvenet

Differential Revision: D6673436

fbshipit-source-id: 44f3b51c8c0889c8a9fa652789437970d557fc4c