Handle overridden method when merging interface methods
Summary:
As the result of the investigation of a few recent errors on FB4A, we found that we don't properly handle default implementation of interface method resided in the base type when merging.
This is a serious part that is missing from our transformation.
The change fixes this bug based on the current structure of the analysis.
If we were to support more complex hierarchies, we need to go through a bigger change.

Reviewed By: emma0303

Differential Revision: D6735042

fbshipit-source-id: fb892e7267f17c9461f84c69ccf15a248141617d