Do different things after regalloc
Summary:
CopyProp runs before and after Regalloc. So, we only need to do the extra careful checks after regalloc. Add a field to PassManager to keep track of whether or not regalloc has run yet.

This is kinda messy. Maybe in the long run we should split out CopyProp into two different passes. One for before regalloc (with all the bells and whistle) and a stripped down version after regalloc that only removes redundant moves that regalloc left behind.

Reviewed By: arnaudvenet

Differential Revision: D6750955

fbshipit-source-id: b3078119990b3ff391560f437820a9a5c2a70b39