Allow for dynamic assertion of interprocedural constprop facts
Summary: Add an option for inserting code at the start of each method which checks that the arguments that our static analysis thinks are constant actually have those values at runtime. If the check fails, the code will call out to a handler method, passing it a single integer that indicates the index of the failing parameter. Useful for debugging the pass.

Reviewed By: justinjhendrick, arnaudvenet

Differential Revision: D6694879

fbshipit-source-id: 556a711faf18e299ebb8b6ff9e2d9daf96b6a3fc