Update the keep rule modifiers for easier parallelization
Summary:
First, I slightly updated the name: `blanket_keep` to `blanket_keepnames` for clarification.

I found the potential races in `apply_keep_modifiers` while parallelizing. This patch makes it easy to parallelize.

 ---

Some background:

- Redex implements two keep modifiers:
  `-keep,allowshrinking`  and `-keep,allowobfuscation`.
- Unfortunately, multiple keep rules might have conflict modifiers.

For example, say we have two keep rules, `A` and `B`. And we have a class (or field, method) that is matched to both `A` and `B` rules.

Here, `A` is just a normal keep rule, which means neither shrink nor obfuscate. But, `B` is`-keep,allowshrinking`. What should we do in this case?

It'd be best if we completely avoid such conflicting rules, but programmers must verify its integrity, which isn't easy.

Instead, we just take a conservative approach: in a conflicting case, keep has a priority: don't shrink or obfuscate.

The previous code already implemented this policy: we rejected setting the modifiers once a keep rule has been applied. But the implementation can suffer from data races when running in parallel. This patch converts a single `bool` into a pair of `set` and `unset` booleans. If `unset` is `true`, the modifier is always false.

Reviewed By: justinjhendrick

Differential Revision: D6768768

fbshipit-source-id: 9815a707a9d156dac91407330a8b693a298f0d27