Unreachable invokes should not cause callee to be marked unreachable
Summary: We were using an Environment instead of a Partition to represent our collection of program states in IPC. This resulted in a bug: a single unreachable invoke instruction in a method would cause the entire environment to be set to Bottom, which meant that all the invokes in that method would be incorrectly marked as unreachable. This caused us to wrongly infer values to be constant when they weren't.

Reviewed By: arnaudvenet

Differential Revision: D6793066

fbshipit-source-id: 7fbfa1b462cf70f70eefd7e8c576fc332f30064f