IPC: Add static field analysis
Summary:
If a field only ever gets written to by constant values, we can treat it as a constant in our analysis. This diff does that for static fields (instance fields require more complicated analysis).

I do think the current implementation is a bit messy; the division between interprocedural and intraprocedural functionality is not very clean at the moment. I plan to follow up this diff with a refactoring / reorganization of constprop.

Reviewed By: arnaudvenet

Differential Revision: D6827457

fbshipit-source-id: 66596f5572b735998bba26fef66f2c05583bab1b