Don't clear the ids to avoid before obfuscating private members
Summary:
This is a workaround for a potential bug I found with the next diff. What this
change does is not allowing private fields to be renamed to "old" public field
names. (old since they were renamed)
Since normally, field names shouldn't be used for renaming, this shouldn't
affect most of the cases.

There are different things that I think can be done for this optimization:
* don't rename if the renamed length is > than the old one
* when actually updating the field spec, we need to preserve the order in which field names were changed (this would be the actual fix, instead of this proposed workaround)
* reduce nondeterminism (now it is pretty hard to understand what changes because of the non-determinism :( )

The problem:
* public field (1) class.<old_public_name> -> gets renamed to class.<random_name>
* private field (2) class.<old_private_name> -> gets renamed to class.<old_public_name>
Tehnically speaking, this is valid (names don't collide), but when updating the
field specs we should preserve the order: update (1) and then (2)

Reviewed By: dariorussi

Differential Revision: D6843849

fbshipit-source-id: 65de5a91a4771653e31dd2e98aac3f9b35a8a030