create IRList
Summary:
Rename `FatMethod` to `IRList`. This is part of the overall effort to create an editable `CFG` as the source of truth.

In preparation for truly editing the Control Flow Graph, I've decoupled `IRList` (née `FatMethod`) and `IRCode`. The reason they need to be decoupled is that each `Block` in the `CFG` has an `IRList`, and if we want to edit those we need to have methods that operate on `IRList`s, not `IRCode`s.

* `IRList` encapsulates boost's intrusive list (instead of just a `using` declaration)
  * Soon we'll add and change the methods on here
* `IRCode` still has the Dex syncing code (because that's specific to `DexMethod`, and not generally useful elsewhere, like in `CFG`)
* In the short term (while `CFG` still exist in non-editable mode), `IRCode` will act as a switch depending on the current mode. Optimizations will either edit the `CFG` or the linear representation. `IRCode` will have the `if else` branch to handle the mode.
  * In the long term, `IRCode` will create the `CFG` immediately after we convert from `DexCode`.
  * Or maybe `IRCode` will be deleted entirely

The important changes are in the following files:
* IRList.cpp/h
* IRCode.cpp/h
* ControlFlow.cpp/h

Almost everything else is just the rename from `FatMethod` to `IRList`.

Reviewed By: int3, arnaudvenet

Differential Revision: D6827522

fbshipit-source-id: f43060f08ff89a4a83ae5e8e7519e2da7335c81c