regalloc: Fix range alignment logic
Summary:
I intended for the allocator to try and align range instructions with similar
register operand sequences. However, in practice this did not actually happen
-- given two alignable range instructions:

  invoke-static (v0 v1) LFoo;.bar:(II)V
  invoke-static (v0 v1) LFoo;.baz:(II)V

we would insert move instructions for v0 and v1 when processing the call to
Foo.baz(), because our criterion for spilling was simply to check if each given
symreg was already mapped to a vreg. We neglected to check if the vreg it was
mapped to could be reused. (Well we did have that check, but only in one out
of the 3 places where it was needed.)

Reviewed By: emma0303

Differential Revision: D6830045

fbshipit-source-id: 85343371f9a5ea9feda25f865d5065ac84d117a9