regalloc: Avoid spilling reg operands of single-arg invokes
Summary:
Another case of code duplication leading to bugs. There were two similar pieces
of logic to determine the maximum allowable vreg for a src symreg, but one of
them forgot to handle the special case of single-arg invokes. As a result we
would insert spill code for them if some other instruction shared the same
symreg operand and needed a spill.

Reviewed By: arnaudvenet

Differential Revision: D6830131

fbshipit-source-id: 34fd54197c544d8b10e1b553047ba7b614126a45