regalloc: Don't insert unnecessary spills for aliased ranges
Summary:
Previously, we would record which range symregs needed spilling and then insert
one instruction whenever we saw that symreg in the register operand list. This
meant that for an invoke-range instruction with a pair of aliased registers --
like

  invoke-static (v0 v0 v1 v2 v3 v4) ...

-- we would insert two spill instructions, one for each occurrence of v0, even
if only one of the two operands needed to be spilled.

This diff changes things so that instead of recording the symregs, we record
the indices of the operands that need spilling. So in the example above, we are
now able to spill only one out of the 2 `v0` operands (if our selection
heuristics deem that to be optimal).

Reviewed By: arnaudvenet

Differential Revision: D6830296

fbshipit-source-id: b3d0e5c094060d52515aaedf523c48b7c3677c5d